Fix this in CommonJS require
In modules included by `require`, `this` should be an empty object, but
instead it was set to the same as the scope.

To fix this, we need to add a new parameter to the Script.exec() function so that "scope" and "this" can both be specified. The old function has been deprecated.