Start buliding `MapShape`s for the interpreter, but not using yet.
Experimented with building `MapShape`s in various places. The obvious
place felt like it would be in the ScriptNode symbol flattening code,
but that is called too often, and generates maps during tests of over
7000 entries! So, we need a builder for generating these maps which
will not create a new map at every step (because those intermediate
maps will never be accessible).

I then found that we really need to support maps with duplicate names
because of things like

```javascript
function f(a, a, a) { return arguments; }
```

So even though `a` when accessed by name will only give us the third
argument we don't want to get rid of those internal offsets. Detecting
and rejecting duplicates also fails currently because of other bugs,
so for now I'm just allowing these.