Refactor to always use shape for local vars in interpreter.
Used this to tease out any issues with shapes vs. the old argNames and
attributes arrays, and there were a couple. But we now have an
interpreter which no longer needs the args array, but we don't yet
have a `NativeCall` object which is making use of the shape. Rather
than trying to do anything fancy with an invokeDynamic call for
compiled functions I'm instead going to introduce a `getShape()` method
and store the shape itself in a static. I'll remove the new shape
argument from all the places I added it, and get it via this new method.